Crate tamawiki

A wiki implemented in Rust

Modules

connection
Handles websocket connections to collaborative edit sessions

document
This module describes a Document, which represents some String content at a point in time, and Operations, which can be applied to the Document to update it's content.

store
A store persists document updates using paths as keys. Every store implementation must define an Actor capable of handling the messages in this module by implementing the Store trait.

templates
Loading and rendering of Tera templates

Structs

TamaWikiState
Per-thread application state

Functions

app
Creates a new TamaWiki actix_web App

server
Creates a new TamaWiki HTTP server and binds to the given address