Module tamawiki::connection

Client communication with an EditSession via websockets

Structs

Change
Broadcast by the server to all participants, except the author, when it accepts a new change to the document.

Connected
Sent by the server when the client has been successfully connected to an EditSession

Connection
Represents a client connection to an edit session

Edit
Sent by the client when local changes were made to their document

Join
Sent by the server when a participant joins the EditSession

Leave
Sent by the server when a participant leaves the EditSession

ServerError
An error message sent by the server to the client

Enums

ClientMessage
An enum of the possible messages the client can send to the server

ServerMessage
An enum of the possible messages the server can send to the client

Type Definitions

ConnectionId
Connection identifier, which must be unique for the current edit session