Module tamawiki::document

Documents and Operations on their content.

Structs

Delete
Deletes a contiguous region of content from the Document

Document
Represents some String content at a point in time

Edit
An Edit combines multiple operations into a single Document change (i.e. all the operations are applied together, or not at all).

Insert
Inserts new content at a single position in the Document

Join
A new participant has joined the DocumentSession

Leave
A participant has left the DocumentSession

Enums

EditError
Error conditions which may occur when applying an Operation to a Document.

Event
Describes an event corresponding to a single Document.

Operation
Describes incremental changes to a Document's content. Through the accumulated application of Operations to a Document, the Document's content at a point in time can be derived.

Type Definitions

ParticipantId
Connection identifier, which must be unique between concurrent events