Module tamawiki::store

Persists document updates using paths as keys.

Every store backend must implement the Store trait.

A document's content is obtained by the accumulated application of updates within the store up until a given point in time. A store implementation may checkpoint or cache these update applications in order to speed up this process.

Modules

memory
An in-memory store, useful for testing.

Enums

StoreError
Error conditions for reading data from, or writing data to, the store.

Traits

Store
Every store backend must provide a client interface to the store using the Store trait. This client interface will be cloned for multiple requests and orchestrating concurrent access (e.g. via locking and a connection pool) is left to the implementation.

Type Definitions

SequenceId
The sequence number for an Update. The first Update for a Document is SequenceId=1 (not 0). This is so requesting updates since SequenceId=0 will return all Updates for a Document.