Trait tamawiki::store::Store

pub trait Store: Clone + Debug + Send + 'static {
    type Stream: Stream<Item = (SequenceId, Event), Error = StoreError> + Send;
    type SinceFuture: Future<Item = Self::Stream, Error = StoreError> + Send;
    type PushFuture: Future<Item = SequenceId, Error = StoreError> + Send;
    fn push(&mut self, path: PathBuf, event: Event) -> Self::PushFuture;

    fn seq(
        &self, 
        path: &Path
    ) -> Box<Future<Item = SequenceId, Error = StoreError> + Send>;

    fn since(&self, path: &Path, seq: SequenceId) -> Self::SinceFuture;

    fn content(
        &self, 
        path: &Path
    ) -> Box<Future<Item = (SequenceId, Document), Error = StoreError> + Send>;

    fn content_at(
        &self, 
        path: &Path, 
        seq: SequenceId
    ) -> Box<Future<Item = Document, Error = StoreError> + Send>;
}

Every store backend must provide a client interface to the store using the Store trait. This client interface will be cloned for multiple requests and orchestrating concurrent access (e.g. via locking and a connection pool) is left to the implementation.

Associated Types

type Stream: Stream<Item = (SequenceId, Event), Error = StoreError> + Send

Type for the stream of Events returned by since() calls

type SinceFuture: Future<Item = Self::Stream, Error = StoreError> + Send

The Future returned by since() calls

type PushFuture: Future<Item = SequenceId, Error = StoreError> + Send

The Future returned by push() calls

Required Methods

fn push(&mut self, path: PathBuf, event: Event) -> Self::PushFuture

Adds a new Event to the document at 'path' and returns the new SequenceId. If the document does not exist, the act of pushing an Event creates it.

fn seq(
&self,
path: &Path
) -> Box<Future<Item = SequenceId, Error = StoreError> + Send>

Requests the current SequenceId for the document at 'path', or StoreError::NotFound if it does not exist.

fn since(&self, path: &Path, seq: SequenceId) -> Self::SinceFuture

Requests a stream of Events starting after the provided SequenceId. Requesting the current (head) SequenceId is not an error, but will return an empty stream. Requesting Events since a SequenceId that does not exist yet is a StoreError::InvalidSequenceId.

fn content(
&self,
path: &Path
) -> Box<Future<Item = (SequenceId, Document), Error = StoreError> + Send>

Requests the current SequenceId and content for the document at 'path' (with all events applied), or StoreError::NotFound if the document does not exist.

fn content_at(
&self,
path: &Path,
seq: SequenceId
) -> Box<Future<Item = Document, Error = StoreError> + Send>

Requests a snapshot of the document's content at a specific SequenceId. All events from SequenceId=1 (inclusive) to SequenceId=seq will be applied. Results in a StoreError::NotFound if the document does not exist, and StoreError::InvalidSequenceId if the SequenceId does not exist.